adding form image upload option changed create-project.php changed app.jsx moved create project form changed default state to catalog
